rename HAL to HC32 - rename HAL/HC32F46x -> HAL/HC32 - rename pins/hc32f46x -> pins/hc32f460 - replace TARGET_HC32F46x -> ARDUINO_ARCH_HC32 - move HC32 board definition - rename pio env
