+++ Marlin/src/HAL/HC32/HAL.cpp
