+++ Marlin/src/HAL/HC32/MarlinSerial.cpp
@@ -20,7 +20,7 @@
  *
  */
 
-#ifdef TARGET_HC32F46x
+#ifdef ARDUINO_ARCH_HC32
 #include "MarlinSerial.h"
 #include "../../inc/MarlinConfig.h"
 #include <drivers/usart/Usart.h>
@@ -137,4 +137,4 @@ CHECK_AXIS_SERIAL(E6);
 #if AXIS_HAS_HW_SERIAL(E7)
 CHECK_AXIS_SERIAL(E7);
 #endif
-#endif // TARGET_HC32F46x
+#endif // ARDUINO_ARCH_HC32
