+++ Marlin/src/HAL/HC32/README.md
@@ -1,3 +1,3 @@
-HC32F46x HAL based on the sources published by Voxlab at [Voxelab-64/Aquila_X2](https://github.com/Voxelab-64/Aquila_X2), [Version 5.1.8](https://github.com/Voxelab-64/Aquila_X2/commit/d1f23adf96920996b979bc31023d1dce236d05db).
+HC32F460 HAL based on the sources published by Voxlab at [Voxelab-64/Aquila_X2](https://github.com/Voxelab-64/Aquila_X2), [Version 5.1.8](https://github.com/Voxelab-64/Aquila_X2/commit/d1f23adf96920996b979bc31023d1dce236d05db).
 
 Some of the changes done in [alexqzd/Marlin-H32](https://github.com/alexqzd/Marlin-H32) have been incorperated aswell, based on the state at [commit 404a201](https://github.com/alexqzd/Marlin-H32/tree/404a201487af58a959edf0af508211dc047c2f3b)
