+++ Marlin/src/HAL/HC32/dogm/u8g_com_stm32duino_fsmc.cpp
@@ -28,8 +28,8 @@
 
 #include "../../../inc/MarlinConfig.h"
 
-#if defined(TARGET_HC32F46x) && PIN_EXISTS(FSMC_CS) // FSMC on 100/144 pins SoCs
-#warning "'u8g_com_stm32duino_fsmc' has not been adapted to HC32F46x. Proceed at your own risk"
+#if defined(ARDUINO_ARCH_HC32) && PIN_EXISTS(FSMC_CS) // FSMC on 100/144 pins SoCs
+#warning "'u8g_com_stm32duino_fsmc' has not been adapted to HC32F460. Proceed at your own risk"
 
 #if HAS_MARLINUI_U8GLIB
 
@@ -425,4 +425,4 @@ void LCD_IO_WaitSequence_Async()
 #endif // LCD_USE_DMA_FSMC
 
 #endif // HAS_MARLINUI_U8GLIB
-#endif // TARGET_HC32F46x && FSMC_CS_PIN
+#endif // ARDUINO_ARCH_HC32 && FSMC_CS_PIN
