+++ Marlin/src/HAL/HC32/eeprom_sdcard.cpp
@@ -24,7 +24,7 @@
  * Implementation of EEPROM settings in SD Card
  */
 
-#ifdef TARGET_HC32F46x
+#ifdef ARDUINO_ARCH_HC32
 
 #include "../../inc/MarlinConfig.h"
 
@@ -120,4 +120,4 @@ bool PersistentStore::read_data(int &pos, uint8_t *value, const size_t size, uin
 }
 
 #endif // SDCARD_EEPROM_EMULATION
-#endif // TARGET_HC32F46x
+#endif // ARDUINO_ARCH_HC32
