+++ Marlin/src/HAL/HC32/endstop_interrupts.cpp
@@ -67,7 +67,7 @@ void setup_endstop_interrupts()
 // if you encounter this error, you'll have to disable some endstops
 #define ENDSTOPS_INTERRUPTS_COUNT COUNT_ENABLED(USE_X_MAX, USE_X_MIN, USE_Y_MAX, USE_Y_MIN, USE_Z_MAX, USE_Z_MIN, USE_Z2_MAX, USE_Z2_MIN, USE_Z3_MAX, USE_Z3_MIN, USE_Z_MIN_PROBE)
 #if ENDSTOPS_INTERRUPTS_COUNT > 10
-  #error "too many endstop interrupts! HC32F46x only supports 10 endstop interrupts."
+  #error "too many endstop interrupts! HC32F460 only supports 10 endstop interrupts."
 #endif
 #if ENDSTOPS_INTERRUPTS_COUNT == 0
   #error "no endstop interrupts are enabled! If you wish to ignore this error, please comment out this line."
