+++ Marlin/src/HAL/HC32/pinsDebug.h
