+++ Marlin/src/HAL/platforms.h
@@ -35,8 +35,8 @@
   #define HAL_PATH(PATH, NAME) XSTR(PATH/HAL/TEENSY40_41/NAME)
 #elif defined(TARGET_LPC1768)
   #define HAL_PATH(PATH, NAME) XSTR(PATH/HAL/LPC1768/NAME)
-#elif defined(TARGET_HC32F46x)
-  #define HAL_PATH(PATH, NAME) XSTR(PATH/HAL/HC32F46x/NAME)
+#elif defined(ARDUINO_ARCH_HC32)
+  #define HAL_PATH(PATH, NAME) XSTR(PATH/HAL/HC32/NAME)
 #elif defined(__STM32F1__) || defined(TARGET_STM32F1)
   #define HAL_PATH(PATH, NAME) XSTR(PATH/HAL/STM32F1/NAME)
 #elif defined(ARDUINO_ARCH_STM32)
