+++ Marlin/src/libs/BL24CXX.cpp
@@ -48,7 +48,7 @@
 #if defined(__STM32F1__)
   #define SDA_IN()  do{ PIN_MAP[IIC_EEPROM_SDA].gpio_device->regs->CRH &= 0XFFFF0FFF; PIN_MAP[IIC_EEPROM_SDA].gpio_device->regs->CRH |= 8 << 12; }while(0)
   #define SDA_OUT() do{ PIN_MAP[IIC_EEPROM_SDA].gpio_device->regs->CRH &= 0XFFFF0FFF; PIN_MAP[IIC_EEPROM_SDA].gpio_device->regs->CRH |= 3 << 12; }while(0)
-#elif defined(STM32F1) || defined(STM32F4) || defined(TARGET_HC32F46x)
+#elif defined(STM32F1) || defined(STM32F4) || defined(ARDUINO_ARCH_HC32)
   #define SDA_IN()  SET_INPUT(IIC_EEPROM_SDA)
   #define SDA_OUT() SET_OUTPUT(IIC_EEPROM_SDA)
 #endif
