+++ ini/hc32.ini
@@ -5,7 +5,7 @@
 
 #################################
 #
-# HC32F46x Architecture with unified HC32 HAL 
+# HC32F460 Architecture with unified HC32 HAL 
 # 
 # Naming Example: HC32F460JEUA
 #
@@ -30,7 +30,7 @@
 platform = https://github.com/shadow578/platform-hc32f46x.git
 board = generic_hc32f460
 build_flags = 
-	-D TARGET_HC32F46x 							# marlin target
+	-D ARDUINO_ARCH_HC32
 	-D REDIRECT_PRINTF_TO_SERIAL 				# redirect core-provided printf to host serial
 	-D __DEBUG									# DDL debug mode
 	-D __CORE_DEBUG								# core debug mode
@@ -65,22 +65,22 @@ board_build.flags.cpp =
 #
 # HC32F460xCxx (256K Flash)
 #
-[common_HC32F460xCxx_variant]
+[common_HC32F460C_variant]
 extends = common_HC32_variant
 board_build.ld_args.flash_size = 256K
 
 #
 # HC32F460xExx (512K Flash)
 #
-[common_HC32F460xExx_variant]
+[common_HC32F460E_variant]
 extends = common_HC32_variant
 board_build.ld_args.flash_size = 512K
 
 #
-# Voxelab Aquila V1.0.1 Mainboard, as found in the Voxelab Aquila X2
+# Aquila V1.0.1 Mainboard, as found in the Voxelab Aquila X2
 #
-[env:HC32F460xCxx_AQUILA_V101]
-extends = common_HC32F460xCxx_variant
+[env:HC32F460C_AQUILA_V101]
+extends = common_HC32F460C_variant
 # bootloader start address, as logged by the bootloader on boot
 board_build.ld_args.flash_start = 0xC000
 
@@ -90,5 +90,5 @@ board_build.ld_args.boot_mode = secondary
 
 # print panic messages to Serial2 (host serial)
 build_flags = 
-	${common_HC32F460xCxx_variant.build_flags}	# extend common flags
+	${common_HC32F460C_variant.build_flags}	# extend common flags
 	-D PANIC_USART2_TX_PIN=PA9					# write panic messages to PA9 (TX of host serial) using USART2
